// types/member.ts
export interface Member {
  _id: string;
  fullName: string;
  sex: string;
  age: number;
  phoneNumber: string;
  studentId: string;
  college: string;
  department: string;
  section: string;
  academicYear: string;
  membershipStatus: string;
  fellowshipTeam: string;
  leadershipRole?: string;
  attendingBibleStudy: string;
  bibleStudyRole?: string;
  bornAgain: string;
  churchName: string;
  spiritualGift: string;
  favoriteBibleVerse: string;
  prayerRequest?: string;
  createdAt: string;
}